pyproject.toml
[tool.poetry]
name = "rest-crm"
version = "0.1.0"
description = ""
authors = ["Bogdan Vyatkin"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.1"
asgiref = "3.6.0"
cffi = "1.15.1"
cryptography = "40.0.2"
djangorestframework = "3.14.0"
djangorestframework-jwt = "1.11.0"
pillow = "9.5.0"
pycparser = "2.21"
pyjwt = "1.7.1"
pytz = "2023.3"
sqlparse = "0.4.4"
django-rest-swagger = "2.2.0"
drf-yasg = "^1.21.5"
psycopg2-binary = "2.9.6"
django-cors-headers = "^4.0.0"
django-prometheus = "^2.3.1"
prometheus-client = "^0.17.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

